Latest push, more logs and some captcha

+++ src/com/servlet/MainController.java
@@ -143,6 +143,7 @@ public class MainController{
 		System.out.println("im in");
 		boolean out = false;
 		
+		
 		if(request.getSession().getAttribute("userType")!=null) {
 			int type = Integer.parseInt(request.getSession().getAttribute("userType").toString());
 			
@@ -207,6 +208,7 @@ public class MainController{
 			System.out.println("ReserveBook");
 			out = BookService.reserveBook(idbook);
 			transaction = TransactionService.addTransaction(t);
+			
 		}
 		else
 			request.getRequestDispatcher("WEB-INF/LoginPage.jsp").forward(request, response);
@@ -517,6 +519,9 @@ public class MainController{
 						if(UserService.addUser(newUser)){
 							status.setSucess(true);
 							System.out.println("Added");
+							LogsService.logAction(	"AddEmployee", 
+													"[Added Employee] Employee: "+newUser.getEmail()+":"+newUser.getId()+"'Added "+ request.getRemoteAddr(),
+													(Integer)request.getSession().getAttribute("userID"));
 						}else
 							status.setMessage("Check the ID Number or Email, Account might exist already");
 					}else
@@ -743,6 +748,8 @@ public class MainController{
 				System.out.println(employeeJSON);
 				request.setAttribute("employeeJSON", employeeJSON);
 				request.getRequestDispatcher("EmployeeEdit.jsp").forward(request, response);
+				
+
 			}
 			else
 				request.getRequestDispatcher("AccessDenied.jsp").forward(request, response);
@@ -755,7 +762,9 @@ public class MainController{
 	@ResponseBody
 	public String editedEmployee(@ModelAttribute("employee") User employee, @RequestParam("isDelete") boolean isDelete,
 									@RequestParam("isAnswerChanged") boolean isAnswerChanged, @RequestParam("isQuestionChanged") boolean isQuestionChanged,
-									@RequestParam("isPasswordChanged") boolean isPasswordChanged, HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+									@RequestParam("isPasswordChanged") boolean isPasswordChanged,
+									@RequestParam("grecaptcharesponse") String cap,
+									HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		System.out.println("EDITING EMPLOYEE");
 		boolean success = false;
 		if(request.getSession().getAttribute("userType")!=null) {
@@ -764,28 +773,38 @@ public class MainController{
 			if(AuthorityCheckerService.isAdmin(type)) {
 				EncryptionService encode = new EncryptionService();
 				
-				if(!isDelete) {
-					success = UserService.editEmployee(employee);
-					
-					if(isPasswordChanged) {
-						System.out.println("CHANGED PASSWORD");
-						String password = encode.encryptPass(employee.getPassword());
-						success = success && UserService.passwordChange(password, employee.getId());
+				if(!this.processCaptcha(request, cap)){
+					if(!isDelete) {
+						success = UserService.editEmployee(employee);
+						LogsService.logAction("EditEmloyee", "[Edit Employee] Employee: "+employee.getEmail()+":"+employee.getId()+" was changed. "+ request.getRemoteAddr(), (Integer)request.getSession().getAttribute("userID"));
+
+						if(isPasswordChanged) {
+							System.out.println("CHANGED PASSWORD");
+							String password = encode.encryptPass(employee.getPassword());
+							success = success && UserService.passwordChange(password, employee.getId());
+
+							LogsService.logAction("EditEmloyee", "[Edit Employee Pass Change] Employee: "+employee.getEmail()+":"+employee.getId()+"'s Password was changed. "+ request.getRemoteAddr(), (Integer)request.getSession().getAttribute("userID"));
+						}
+						if(isQuestionChanged) {
+							System.out.println("CHANGED QUESTION");
+							success = success && UserService.questionChange(employee.getSecretQuestion(), employee.getId());
+							LogsService.logAction("EditEmloyee", "[Edit Employee Question Change] Employee: "+employee.getEmail()+":"+employee.getId()+"'s Secret Question was changed. "+ request.getRemoteAddr(), (Integer)request.getSession().getAttribute("userID"));
+						}
+						if(isAnswerChanged) {
+							System.out.println("CHANGED ANSWER");
+							String answer = encode.encryptPass(employee.getSecretAnswer());
+							success = success && UserService.answerChange(answer, employee.getId());
+
+							LogsService.logAction("EditEmloyee", "[Edit Employee Secret Answer Changed] Employee: "+employee.getEmail()+":"+employee.getId()+"'s Secret Answer was changed. "+ request.getRemoteAddr(), (Integer)request.getSession().getAttribute("userID"));
+						}
 					}
-					if(isQuestionChanged) {
-						System.out.println("CHANGED QUESTION");
-						success = success && UserService.questionChange(employee.getSecretQuestion(), employee.getId());
+					else {
+						System.out.println("DELETING EMPLOYEE");
+						System.out.println(employee.getId());
+						success = UserService.deleteEmployee(employee.getId());
+						LogsService.logAction("EditEmloyee", "[Employee Deleted] Employee: "+employee.getEmail()+":"+employee.getId()+" was Deleted. "+ request.getRemoteAddr(), (Integer)request.getSession().getAttribute("userID"));
+
 					}
-					if(isAnswerChanged) {
-						System.out.println("CHANGED ANSWER");
-						String answer = encode.encryptPass(employee.getSecretAnswer());
-						success = success && UserService.answerChange(answer, employee.getId());
-					}
-				}
-				else {
-					System.out.println("DELETING EMPLOYEE");
-					System.out.println(employee.getId());
-					success = UserService.deleteEmployee(employee.getId());
 				}
 			}
 			else
@@ -796,13 +815,14 @@ public class MainController{
 		
 		return ""+success;
 	}
-//	@RequestMapping(value="/EditProduct", method = RequestMethod.POST)
-//	public void editProduct(HttpServletRequest request, HttpServletResponse response)throws ServletException, IOException{
-//		System.out.println("EditProduct");
-//
-//		request.getSession().setAttribute("productID", request.getParameter("bookID"));
-//
-//	
-//	}
+	
+	@RequestMapping(value="/EditProduct", method = RequestMethod.POST)
+	public void editProduct(HttpServletRequest request, HttpServletResponse response)throws ServletException, IOException{
+		System.out.println("EditProduct");
+
+		request.getSession().setAttribute("productID", request.getParameter("bookID"));
+
+	
+	}
 
 }
